moved `repr_html` code to utils
